bullet-proofing top-level API

+++ AlembicImporterPlugin/AlembicImporter.cpp
@@ -38,47 +38,64 @@ aiCLinkage aiExport aiContext* aiCreateContext(int uid)
 
 aiCLinkage aiExport void aiDestroyContext(aiContext* ctx)
 {
-    aiContext::destroy(ctx);
+    if (ctx)
+    {
+        aiContext::destroy(ctx);
+    }
 }
 
 
 aiCLinkage aiExport bool aiLoad(aiContext* ctx, const char *path)
 {
-    return ctx->load(path);
+    return (ctx ? ctx->load(path) : false);
 }
 
 aiCLinkage aiExport void aiSetConfig(aiContext* ctx, const aiConfig* conf)
 {
-    ctx->setConfig(*conf);
+    if (ctx)
+    {
+        ctx->setConfig(*conf);
+    }
 }
 
 aiCLinkage aiExport float aiGetStartTime(aiContext* ctx)
 {
-    return ctx->getStartTime();
+    return (ctx ? ctx->getStartTime() : 0.0f);
 }
 
 aiCLinkage aiExport float aiGetEndTime(aiContext* ctx)
 {
-    return ctx->getEndTime();
+    return (ctx ? ctx->getEndTime() : 0.0f);
 }
 
 aiCLinkage aiExport aiObject* aiGetTopObject(aiContext* ctx)
 {
-    return ctx->getTopObject();
+    return (ctx ? ctx->getTopObject() : 0);
 }
 
 aiCLinkage aiExport void aiDestroyObject(aiContext* ctx, aiObject* obj)
 {
-    ctx->destroyObject(obj);
+    if (ctx)
+    {
+        ctx->destroyObject(obj);
+    }
 }
 
 aiCLinkage aiExport void aiUpdateSamples(aiContext* ctx, float time)
 {
-    ctx->updateSamples(time);
+    if (ctx)
+    {
+        ctx->updateSamples(time);
+    }
 }
 
 aiCLinkage aiExport void aiEnumerateChild(aiObject *obj, aiNodeEnumerator e, void *userData)
 {
+    if (!obj)
+    {
+        return;
+    }
+    
     size_t n = obj->getNumChildren();
 
     for (size_t i = 0; i < n; ++i)
@@ -97,114 +114,138 @@ aiCLinkage aiExport void aiEnumerateChild(aiObject *obj, aiNodeEnumerator e, voi
 
 aiCLinkage aiExport const char* aiGetNameS(aiObject* obj)
 {
-    return obj->getName();
+    return (obj ? obj->getName() : "");
 }
 
 aiCLinkage aiExport const char* aiGetFullNameS(aiObject* obj)
 {
-    return obj->getFullName();
+    return (obj ? obj->getFullName() : "");
 }
 
 
 aiCLinkage aiExport void aiSchemaSetSampleCallback(aiSchemaBase* schema, aiSampleCallback cb, void* arg)
 {
-    schema->setSampleCallback(cb, arg);
+    if (schema)
+    {
+        schema->setSampleCallback(cb, arg);
+    }
 }
 
 aiCLinkage aiExport void aiSchemaSetConfigCallback(aiSchemaBase* schema, aiConfigCallback cb, void* arg)
 {
-    schema->setConfigCallback(cb, arg);
+    if (schema)
+    {
+        schema->setConfigCallback(cb, arg);
+    }
 }
 
 aiCLinkage aiExport const aiSampleBase* aiSchemaUpdateSample(aiSchemaBase* schema, float time)
-{
-    return schema->updateSample(time);
+{    
+    return (schema ? schema->updateSample(time) : 0);
 }
 
 aiCLinkage aiExport const aiSampleBase* aiSchemaGetSample(aiSchemaBase* schema, float time)
 {
-    return schema->findSample(time);
+    return (schema ? schema->findSample(time) : 0);
 }
 
 
 aiCLinkage aiExport bool aiHasXForm(aiObject* obj)
 {
-    return obj->hasXForm();
+    return (obj ? obj->hasXForm() : false);
 }
 
 aiCLinkage aiExport aiXForm* aiGetXForm(aiObject* obj)
 {
-    return &(obj->getXForm());
+    return (obj ? &(obj->getXForm()) : 0);
 }
 
 aiCLinkage aiExport void aiXFormGetData(aiXFormSample* sample, aiXFormData *outData)
 {
-    sample->getData(*outData);
+    if (sample)
+    {
+        sample->getData(*outData);
+    }
 }
 
 
 aiCLinkage aiExport bool aiHasPolyMesh(aiObject* obj)
 {
-    return obj->hasPolyMesh();
+    return (obj ? obj->hasPolyMesh() : false);
 }
 
 aiCLinkage aiExport aiPolyMesh* aiGetPolyMesh(aiObject* obj)
 {
-    return &(obj->getPolyMesh());
+    return (obj ? &(obj->getPolyMesh()) : 0);
 }
 
 aiCLinkage aiExport void aiPolyMeshGetSummary(aiPolyMesh* schema, aiMeshSummary* summary)
 {
-    schema->getSummary(*summary);
+    if (schema)
+    {
+        schema->getSummary(*summary);
+    }
 }
 
 aiCLinkage aiExport void aiPolyMeshGetSampleSummary(aiPolyMeshSample* sample, aiMeshSampleSummary* summary, bool forceRefresh)
 {
-    sample->getSummary(forceRefresh, *summary);
+    if (sample)
+    {
+        sample->getSummary(forceRefresh, *summary);
+    }
 }
 
 aiCLinkage aiExport int aiPolyMeshGetVertexBufferLength(aiPolyMeshSample* sample, int splitIndex)
 {
-    return sample->getVertexBufferLength(splitIndex);
+    return (sample ? sample->getVertexBufferLength(splitIndex) : 0);
 }
 
 aiCLinkage aiExport void aiPolyMeshFillVertexBuffer(aiPolyMeshSample* sample, int splitIndex, aiMeshSampleData* data)
 {
-    sample->fillVertexBuffer(splitIndex, *data);
+    if (sample)
+    {
+        sample->fillVertexBuffer(splitIndex, *data);
+    }
 }
 
 aiCLinkage aiExport int aiPolyMeshPrepareSubmeshes(aiPolyMeshSample* sample, const aiFacesets* facesets)
 {
-    return sample->prepareSubmeshes(*facesets);
+    return (sample ? sample->prepareSubmeshes(*facesets) : 0);
 }
 
 aiCLinkage aiExport int aiPolyMeshGetSplitSubmeshCount(aiPolyMeshSample* sample, int splitIndex)
 {
-    return sample->getSplitSubmeshCount(splitIndex);
+    return (sample ? sample->getSplitSubmeshCount(splitIndex) : 0);
 }
 
 aiCLinkage aiExport bool aiPolyMeshGetNextSubmesh(aiPolyMeshSample* sample, aiSubmeshSummary* summary)
 {
-    return sample->getNextSubmesh(*summary);
+    return (sample ? sample->getNextSubmesh(*summary) : false);
 }
 
 aiCLinkage aiExport void aiPolyMeshFillSubmeshIndices(aiPolyMeshSample* sample, const aiSubmeshSummary* summary, aiSubmeshData* data)
 {
-    sample->fillSubmeshIndices(*summary, *data);
+    if (sample)
+    {
+        sample->fillSubmeshIndices(*summary, *data);
+    }
 }
 
 
 aiCLinkage aiExport bool aiHasCamera(aiObject* obj)
 {
-    return obj->hasCamera();
+    return (obj ? obj->hasCamera() : false);
 }
 
 aiCLinkage aiExport aiCamera* aiGetCamera(aiObject* obj)
 {
-    return &(obj->getCamera());
+    return (obj ? &(obj->getCamera()) : 0);
 }
 
 aiCLinkage aiExport void aiCameraGetData(aiCameraSample* sample, aiCameraData *outData)
 {
-    sample->getData(*outData);
+    if (sample)
+    {
+        sample->getData(*outData);
+    }
 }
